+++ CHANGES.md
@@ -1,12 +1,11 @@
 Change Log
 ==========
 
-#### next release (8.1.18)
-* MagdaReference can use addOrOverrideAspects trait to add or override "terria" aspect of target.
 #### next release (8.1.19)
 
 * Add WMS support for `TIME=current`
 * Only show `TableMixin.legends` if we have rows in dataColumnMajor and mapItems to show 
+* MagdaReference can use addOrOverrideAspects trait to add or override "terria" aspect of target.
 * [The next improvement]
 
 #### 8.1.18