+++ CHANGES.md
@@ -1,13 +1,16 @@
 Change Log
 ==========
 
-#### next release (8.1.19)
+#### next release (8.1.20)
+
+* [The next improvement]
+
+#### 8.1.19
 
 * Add WMS support for `TIME=current`
 * Only show `TableMixin.legends` if we have rows in dataColumnMajor and mapItems to show
 * Add `WebMapServiceCatalogGroup.perLayerLinkedWcs`, this can be used to enable `ExportWebCoverageService` for **all** WMS layers. `item.linkedWcsCoverage` will be set to the WMS layer `Name` if it is defined, layer `Title` otherwise.
 * MagdaReference can use addOrOverrideAspects trait to add or override "terria" aspect of target.
-* [The next improvement]
 
 #### 8.1.18